backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==4.2.16
django-allauth==65.0.2
django-cleanup==9.0.0
django-cors-headers==4.4.0
django-htmx==1.19.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
Faker==30.0.0
google-auth==2.35.0
httplib2==0.22.0
idna==3.10
numpy==2.0.2
oauth2client==4.1.3
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1-modules==0.4.1
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tailwindcss==0.0.1
toposort==1.10
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3


